requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-crispy-forms==2.6
djangorestframework==3.16.1
mysqlclient==2.2.8
packaging==26.0
pillow==12.1.1
pypdfium2==5.6.0
pytesseract==0.3.13
python-dotenv==1.2.2
sqlparse==0.5.5
msal==1.35.1